requirements.txt
fastapi
uvicorn[standard]
openai
pinecone
python-dotenv
pypdf